python/streamlit-crewai/requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
google-generativeai>=0.3.0
memorystack>=1.0.0
python-dotenv>=1.0.0
